requirements.txt
pyspark==3.2.0
delta-spark==2.0.0